+++ ops/testnet/staging/main.tf
@@ -46,13 +46,11 @@ module "router" {
   timeout                  = 180
   environment              = var.environment
   stage                    = var.stage
-  mnemonic                 = var.mnemonic
-  service_config_value     = local.local_router_config
   ingress_cdir_blocks      = ["0.0.0.0/0"]
   ingress_ipv6_cdir_blocks = []
   service_security_groups  = flatten([module.network.allow_all_sg, module.network.ecs_task_sg])
   cert_arn                 = var.certificate_arn_testnet
-  service_config_name      = "NXTP_CONFIG"
+  container_env_vars       = local.router_env_vars
 }
 
 module "router_logdna_lambda_exporter" {
@@ -91,13 +89,11 @@ module "sequencer" {
   timeout                  = 180
   environment              = var.environment
   stage                    = var.stage
-  mnemonic                 = var.mnemonic
-  service_config_value     = local.local_sequencer_config
   ingress_cdir_blocks      = ["0.0.0.0/0"]
   ingress_ipv6_cdir_blocks = []
   service_security_groups  = flatten([module.network.allow_all_sg, module.network.ecs_task_sg])
   cert_arn                 = var.certificate_arn_testnet
-  service_config_name      = "SEQ_CONFIG"
+  container_env_vars       = local.sequencer_env_vars
 }
 
 module "sequencer_logdna_lambda_exporter" {
@@ -111,6 +107,36 @@ module "sequencer_logdna_lambda_exporter" {
   stage           = var.stage
   vpc_id          = module.network.vpc_id
   log_group_arn   = module.sequencer.log_group_arn
+}
+
+module "web3signer" {
+  source                   = "../../modules/service"
+  region                   = var.region
+  zone_id                  = data.aws_route53_zone.primary.zone_id
+  ecs_cluster_sg           = module.network.ecs_task_sg
+  allow_all_sg             = module.network.allow_all_sg
+  execution_role_arn       = data.aws_iam_role.ecr_admin_role.arn
+  cluster_id               = module.ecs.ecs_cluster_id
+  vpc_id                   = module.network.vpc_id
+  private_subnets          = module.network.private_subnets
+  lb_subnets               = module.network.public_subnets
+  docker_image             = "ghcr.io/connext/web3signer:latest"
+  container_family         = "web3signer"
+  health_check_path        = "/upcheck"
+  container_port           = 9000
+  loadbalancer_port        = 80
+  cpu                      = 256
+  memory                   = 512
+  instance_count           = 1
+  timeout                  = 180
+  environment              = var.environment
+  stage                    = var.stage
+  internal_lb              = true
+  ingress_cdir_blocks      = [module.network.vpc_cdir_block]
+  ingress_ipv6_cdir_blocks = []
+  service_security_groups  = flatten([module.network.allow_all_sg, module.network.ecs_task_sg])
+  cert_arn                 = var.certificate_arn_testnet
+  container_env_vars       = local.web3signer_env_vars
 }
 
 module "network" {